edge_benchmarking/requirements.txt
git+https://github.com/Verz1Lka/scrapy-crawlera-fetch.git@dce1099
google-cloud-bigquery
google-cloud-storage
dateparser
scrapinghub
scrapy-zyte-smartproxy